move exceptions to separate package
